build bcpc-bootstrap via Packer
This change is to provide the means to build the initial bcpc-bootstrap
machine image having similar configuration to that of a bare metal
deployment.  The goal is have a functional machine image that mimics
that exact same state after PXE booting a bare metal bcpc-bootstrap node
and undergoing the typical installation procedure.