remove destructive VirtualBox networking setup
Setting up VirtualBox host-only networking for BCPC was destructive in
that it removed all interfaces and assumed that the only interfaces
available were to be used by BCPC.

This incremental changes removes the destructive nature of the host-only
networking setup.  A separate commit will provide a more intelligent
mechanism for creating the only interfaces if it does not already exist.
As a by product, the Windows path for setting up the host-only
interfaces has also been removed.