Changes Cinder RBD driver behavior with RAW image->volume
This patch alters the behavior of the Cinder RBD driver to greatly speed
operations where a user is creating a volume from a RAW image. Normally
the image is always piped through the API, but if the type of the source
image is RAW, it will instead execute rbd clone and rbd flatten
commands against Ceph and bypass the APIs. This is considerably faster
and more efficient than transferring the image using the APIs (in tests,
about 5-6 times faster).

This requires that the Ceph user Cinder is using to access the pools
have read capability to the images pool. I have added another
configuration item as well to make Cinder aware of the name of the
images pool, since that information was not previously available to
Cinder.