Bump up default VM topology/hardware specs.
We've noticed an increased amount of false-positive build
failures recently due to haproxy timeouts and other items
that suggest the default topology/hardware.yml is too
conservative for recent changes to BVE.

This commit bumps up the amount of vCPUs/RAM allocated to
each VM (by default) to avoid false-positive build failures
and to reduce the amount of swapping that occurs during
builds.

With these changes, 1h1w builds ought to still easily work
on a host with quad-core CPU and 32GB of RAM, which is more
obtainable today in 2020 than it was a few years ago when
these settings were first instated.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>