nova-compute: Avoid rendering of invalid config
If cpu_mode is host-(model|passthrough), then cpu_model
should be disregarded; otherwise, one would have requested
a "custom" CPU model or not specified to use the hosts'
configuration.

In fact, `nova` throws an exception during scheduling if
the `nova-compute` specifies `cpu_model` when `cpu_mode`
is one of host-model, host-passthrough:

```
Invalid: A CPU model name should not be set when a host CPU
model is requested
```

Thus, even if the chef-databag specifies this invalid
set of parameters, this commit ensures that the resulting
config will not result in scheduling efforts (by prefering
the host CPU mode configuration item).

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>