ansible: Run common tasks in parallel during build
Each of the current Ansible groups imports two common
roles: `common` and `chef-node`.  By adding a playbook
for the common role itself, we can leverage this fact to
run Ansible tasks in parallel across all the groups,
whereas it was done serially in a group-by-group fashion
previously.

TBD:
  * `bootstraps` need `chef-node`
  * `headnodes` need `common`

This improves build time of a 1h1w by ~9.4% in my lab.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>