Patch the Neutron RBAC membership test
The current Neutron RBAC membership test results a database
query being issued, which itself makes use of a subquery
that is unable to make use of an index.

As the number of RBACs grows in a cloud deployment, the
missed opportunity for using the index becomes very apparent
with some API calls.  Adjust the SQLalchemy query
generation logic such that MySQL will make use of an index
where it did not previously.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>