Upgrade to OpenStack Stein and python3
Note that watcher is basically broken in Stein as it
stands; we don't intend to support Stein for more than
an upgrade cycle, so that's fine.

Also, noting that `qemu-system-x86` recommends
`qemu-system-gui` starting on Stein+, so stop pulling
in recommends for these packages and installs the bits
we really need piecemeal.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>