Make ProxySQL user updates more reliable
Previously a flag was set in the ProxySQL recipe when user attributes
needed to be updated, and read by subsequent recipes. This would require
a full chef run be completed every time a user attribute was changed,
otherwise subsequent chef runs would not have the said flag set. This
change makes it so that each recipe checks whether its ProxySQL user
needs to be updated.

Signed-off-by: Mike Boruta <mboruta1@bloomberg.net>