neutron/model_query: Fix --share/--no-share filter
When splitting the DB query from the OUTER JOIN into the
two components that are UNION'd late, the 'is_shared'
filter got lost when the admin queries against network-
related collections. This results in networks, subnets,
etc. being listed unconditionally when an admin queries
for items, regardless of whether `--share` or `--no-share`
is specified.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>