Prevent installation of Cinc or Chef >= 15
By default, due to how `apt` works and the packages present
in the Focal repos, `apt` will try to upgrade us to a
Canonical Cinc package aptly named "chef".

We're trying to move away from Chef and just want to keep
the status quo, so pin us to Chef 14.x.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>