Fix server rebuild case for ceph-mon migration
There's yet another (copy/pasted?) block of code used to
source ceph-mon addresses from the monmap that needs
patching when trying to override the monitor IPs to be used
with libvirt domains.

Similarly, found another bug where some monitor addresses
would be truncated from the list due to the use of zip()
when the length of the hosts and ports used for the volumes
was not in equilibrium. Fix this by always just assuming
that all RBD volumes will use the same port and expand the
ports array inline with the monitor hosts array.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>