`neutron-server`: Avoid spurious build failures
Under Focal, `neutron-server` is observed to sometimes
fail to start cleanly the first few runs (exits with a
non-zero code). In response to this, `systemd` attempts to
the process itself a few times before signaling an error
upwards.

Unfortunately, during some builds, it is observed that
start/retry burst limits (configurable via the [Unit]
StartLimitIntervalSec and StartLimitBurst parameters)
are being violated, which ultimately causes `systemd` to
raise an error upwards, and which subsequently causes
`chef-client` (and thus the builds) to fail.

Thus, bump the restart restart interval upwards and
double the default burst limits (as StartLimitIntervalSec
must be > StartLimitBurst * RestartSec to avoid restarting
the service indefinitely in the event of a true failure).

This improves the success rate of Focal builds.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>