configure-node: Breakout system/hardware config
configure-system was getting a little long in the tooth;
try to decompose it into system and hardware tasks.

As part of this, I also realized we no longer need to
consult `sysfs` to see if the I/O scheduler is as-intended.
Before, we set the I/O scheduler via the kernel command
line, and so the change may not have applied yet. But,
since we now use `udev`, we are guaranteed that the change
is rendered out when the `udev` rules are reloaded.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>