cinder: Major recipe refactoring
Until now, we witnessed sporadic CI failures, which were
usually due to some issues with `cinder`. As it turned out,
the CI failures were the canary in the coal mine that
pointed out severe issues with respect to the ordering and
inter-dependencies of resources within the `cinder` recipe.

This commit performs significant overhaul to the recipe
in order to avoid issues with cluster bootstrapping by
reordering resources such that configuration is laid down
before services are started.

The proper ordering of resources means that we can also
eliminate a few service bounces and 'immediate' restarts
that were formerly in place to try and hammer everything
together.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>