Build > 1 OS release worth of Packer boxes
As we prepare to transition into Ubuntu Jammy and split the
management of compute and storage infrastructure, it is
helpful to provide developers with the ability to build
clusters with varying compositions of operating system
releases (e.g., a Focal-fueled OpenStack cluster conjoined
with a Jammy-fueled Ceph cluster).

This commit works in that direction by changing the Packer
tooling such that > 1 OS release worth of Packer box builds
are issued without changing the existing toolchain much.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>