Ensure apparmor capabilities patch is present
Upstream bug:
https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1964636

Focal UCA delivers a new libvirt package which adds
apparmor policy containing capabilities that the apparmor
which ships with Focal does not know about, resulting in
a parse error.

Until we can ensure that the upstream packages we consume
contain the "fixed" version of apparmor from focal-updates,
install them out-of-band as needed.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>