Raise osd_memory_target back to 7GiB
5e35e04 lowered osd_memory_target from 7GiB to 6GiB as
memory pressure was observed for the intended hardware
platform after pages were allocated from NUMA-local
nodes only.

This memory pressure, however, turned out to be from the
libc allocator, which was unintentionally linked in the
Ubuntu builds of Ceph. As we can safely consume 7GiB of
memory per OSD again, do so.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>