Update CPU weight formula
To better spread the load between hosts, the CPU weight
formula of the scheduler now takes into account the VM's
number of requested cores and the CPU idle percentage of
the hypervisor.

Computing the free cores as already used by the requesting
guest will result in a negative weight for hosts that will
be overloaded. This will make for a better informed decision.

The CPU idle percentage is plugged in differently for the two
cases given by the sign of the free cores percentage. The
reason is to not grow the weight of an overloaded host when it
is also less idle.

Signed-off-by: Raul Moldovan <rmoldovan3@bloomberg.net>