ceph: Raise osd_memory_target to 8GiB
We historically had the osd_memory_target to 9GiB, but
it resulted in swapping in edge cases and was lowered to
7GiB.

Ceph as packaged with Jammy was linked with the wrong
memory allocator, and made us push this figure down
further (when we did not know why the RSS was ballooning
to the extent that it was).

Instead of just reverting the bad commit that dropped
this to 7GiB, I realized it probably could be 8GiB based
on metrics gleaned from our hosts - so lets go with that.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>