ceph: Remove bluestore_rocksdb_options override
It seems this was the default taken from Mimic or so but
with the number of compaction jobs changed from 2 to 4
(why...?)

Unfortunately, as part of the new RocksDB in Quincy, this
bit us hard in that bluestore_rocksdb_options upstream
now specifies an additional constraint on maximum WAL size
that we did not pick up.

This is because RocksDB column families each have 4x256MiB
buffers, and Ceph now has a lot of column families. In
practice, this led to nearly a ~100GiB WAL that took an
impossibly long time to recover when an OSD started, and
also would create "slowness" issues some days or weeks
after the OSDs were in production that resulted in them
needing to be turned.

Let's not set these values without some documentation or
explicit need/reason...

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>