Ensure bcpc::nova-head remains idempotent
We already maintain our own version of nova/api-paste.ini
to add the healthchecks in. At the same time, we use
`crudini` to enforce a part of the file with our custom
configuration.

This results in bcpc::nova-head not being idempotent in
some cases, in which case `nova-api` is restarted even
when there are no changes.

To avoid that, make `nova/api-paste.ini` a template and
forward the configuration that we need to control into
the template.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>