Fix bootstrap issue introduced in last commit
HAproxy will not fire up `nova-api` unless the healthchecks
endpoint is up and passing. But, we may have not provisioned
the extended APIs yet and if they are enabled, we'll bounce
`nova-api` configured for a route factory that does not
exist.

To work around this chicken-and-egg problem, use the
upstream Nova's API route factory if we are bootstrapping,
and only enable the extended APIs once we have bootstrapped,
even if the operator enabled them from the get-go.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>