[RDTIBCCOPT-163] System metadata api (#2194)
* add system metadata apis

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* add server optimizations apis

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* rename instance properties APIs

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* minor

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* address comments

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* address more comments

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* more comments

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* comments

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* add missing copyright

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* Fold patches into a hash and loop over them

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Ensure bcpc::nova-head remains idempotent

We already maintain our own version of nova/api-paste.ini
to add the healthchecks in. At the same time, we use
`crudini` to enforce a part of the file with our custom
configuration.

This results in bcpc::nova-head not being idempotent in
some cases, in which case `nova-api` is restarted even
when there are no changes.

To avoid that, make `nova/api-paste.ini` a template and
forward the configuration that we need to control into
the template.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Fix bootstrap issue introduced in last commit

HAproxy will not fire up `nova-api` unless the healthchecks
endpoint is up and passing. But, we may have not provisioned
the extended APIs yet and if they are enabled, we'll bounce
`nova-api` configured for a route factory that does not
exist.

To work around this chicken-and-egg problem, use the
upstream Nova's API route factory if we are bootstrapping,
and only enable the extended APIs once we have bootstrapped,
even if the operator enabled them from the get-go.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Subclass nova/compute/api

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Only allow metadata/properties updates when SHUTOFF

Leverage the subclassed nova/compute/api to ensure that
instance system metadata and properties can only be changed
when the instance is in a SHUTOFF state.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Maintain DB code separately from upstream

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* Fix lint issues

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* lint, schema, and rename

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

* Add missing line to copyright

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>

* address comment

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>

---------

Signed-off-by: Jing Geng <jgeng25@bloomberg.net>
Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>
Co-authored-by: Tyler Stachecki <tstachecki@bloomberg.net>