Add PSI metric to resource tracker
Change the resource monitor for CPU to fetch the pressure stall
information (PSI) from the hypervisor. Along with the other
CPU stats it can be saved as a metric for later use by components
like the scheduler. Such a metric could help relief VM steal time.

As there's no libvirt support for fetching PSI, the way to do it
for now is to perform a read on the kernel interface found at
`/proc/pressure/cpu`. The fetched value represents the 60s average.

See diff here:
https://github.com/raulmldv/nova/commit/b9603b5ca2817131a114a120eafe660c05acf613

Signed-off-by: Raul Moldovan <rmoldovan3@bloomberg.net>