Include PSI in scheduler CPU weight
Use the PSI metric for assiging CPU weights to hosts. Once the
metric is fetched, it can be subtracted from the original formula.
As both the idle time and PSI are time percentages, the result
represents the remaining available time of the CPU. It can be
thought as a measure of overload since it measures the time that
is spent by the threads just fighting on real CPU time.

Signed-off-by: Raul Moldovan <rmoldovan3@bloomberg.net>