RDTIBCC-4728: Fix keystone healthcheck failures
heat.conf declares the heat_domain_admin to be in "Default"
domain, but the user is created by Chef in the "heat"
domain in the Heat recipe.

Change the configuration to reflect this, which ultimately
prevents Heat from trying to lookup a user that does not
exist (thus tripping alarms).

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>