Reinstate bpftool workaround for Felix
This was slated to be fixed in Calico 3.27.0, but the
BPF objects still appear to have not been linked against
a new enough version of libbpf. Thus, ensure that Felix
continues using a version of bpftool that's compatible
with it.

This is slightly less heinous than the former attempt
at a workaround and should allow us to remove it easier
in the future once it is fixed.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>