nova-compute: Operationalize the upstream fix
The upstream fix for disposing of cputune/shares must
only be applied when live-migrating towards a Jammy
hypervisor. If not, we risk sending the instance to a
Focal hypervisor where other domains are running with
1000s of times the number of CPU shares our instance
has. This would result in deathly amounts of CPU steal
under pressure to the instance once it migrates.

We also want to ensure that we still assign CPU shares
when launching on a Focal hypervisor, so restore the
logic which did that.

Finally, we ensure that we recompute the would-be value
for cputune/shares had we never erased it in the case
that we ever migrate from a Jammy-based hypervisor back
to a Focal-based one.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>