Workaround nftables performance regression
calico-felix performs autodetection as to whether or not
to use nftables instead of iptables. On Focal nodes,
nftables was not installed and thus Calico used iptables
on Focal.

For Jammy, nftables is now installed by default and this
auto- detection results in nftables being used. For Ceph
workloads, performance with nftables was comparable to
iptables and as nftables is the forward-looking solution,
we went with that approach. However, the story is very
much different for some compute workloads at the moment,
and quite stark regressions in nftables have been seen.

For this reason, default to using iptables for anything
other than strictly Ceph nodes until its performance is
comparable.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>