Add PKU feature flag when migrating towards Jammy
A recent upstream kernel change [1] addressed a bug wherein
XSAVE stored a PKU feature flag as part of the vCPU context,
but XRSTOR did not load it. A side-effect of this change is
that while VMs which ran on a hypervisor whose host CPU
supported PKU, but which did not expose the same to the
guest, can no longer live-migrate to those same
hypervisors after they said hypervisors are running a
kernel with the aforementioned patch applied.

To permit live-migration in the face of this issue, we use
the Nova live-migration window to add the PKU flag to the
guests as part of live-migrations to kernels which have
effectively have [1] applied to them.

[1] https://lore.kernel.org/lkml/20220205081658.562208-1-leobras@redhat.com/T/

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>